Implementation of a RingBuffer
A Ring Buffer has a fixed capacity. The idea is that adding at the end
(Push) overwrites the oldest entry. The usage is for example for Logs,
History, Chat and so on.

You can consider it as an efficent implementation of "Keep the Last X
in Memory".

Because of its implementation every operation from adding to the start/end
and removing from start/end is O(1).