Removed flatten/bind, corrected append, added roundTrip & equal, changed fold/foldBack
# flatten/bind
flatten/bind didn't make sense in the context of animations,
so i removed it completely.

# append
Previously append runned the first animation to its finish value, always
returned this value and saved the remaining time for the second animation.
But this is not quit right. The finished value is not special and shouldn't
returned, it is just one value of the animation. Instead of returning
the finished value the second animation should be runned with the remaining
time left. This becomes obvious in an animation that runs from a start
to stop and back to start.

# roundTrip
Running something from start to stop and back is probably so common that it
deserves its own function.

# equal
Directly run two animation and check if they produce the same values.
Probably only useful in testing.

# fold/foldBack
Pass the `Anim` DU to `fold` and `foldBack`. This allows for more
transformation functions on Animation if needed.