Removed "key_by" parameter in binary_search
Two functions "key_by" and "comparer" are not needed. Instead of providing
a keying function, the user should select the key in the comparer. This
way we need 1 function call less in the implementation. Giving
it another slight performance boost.

I consider readability the same. It doesn't make the comparer to
complicated.