First Optimization
Actually first optimization was to build all the values that the Draw
commands needs like Rectangle and so on ahead of time instead of building
all the values when the command is called.

But anyway changed the system to draw textures. This is also what later
anyway happens. The system draws small textures. Important for this is
that all particles appear on a single texture atlas, otherwise performance
goes down drastically by losing the performance by batching.

Now it handles ~35.000 particles at 200+ fps. Much better than before.