Change order of operation
first the update. The update updates particles, but also can lead to
deactivated particles. After that Particle.initParticle can directly
use that new freed particles. This leads to a system that exactly
uses the maximum amount of particles per frame. As long creation
of new particles is higher than destruction of particles.