Particle System using structs
Making Particle a struct so we have an Array of Structs instead
of Array of references. It allows for an optimization that elements
don't needed to be swapped anymore.

But makes some things worse as now we need to access a particle
by an index. In practice this change had zero performance benefits
over using a reference-type.