Array of Struct with two Arrays
Same as before but used two Arrays. When an active element is found
then it is written to a previous array. This eliminates the need
to jump in memory and swapping with last element. It can process the
array from start to finish.

Has the disadvantage that structs needs to be copied, but that is also
the only reason why this technique works. Otherwise with reference-types
would be hard to manage to get it correctly working.

Before it was all in one loop. Then separated into two loops. one loop
updates, then writing to previous. Result: Zero fps difference.

Than i added Multi-threading to the updating loop. Also no performance
difference.