Added a Flag to deactivate Rendering of all objects
This way we can see the performance of the Particle System itself, without
that rendering eats up all performance. This is also an example of
"Premature Optimization" and "Micro-benchmarking"

First. All three versions with all rendering on basically have exactly
the same performance. struct_array2 is maybe in most cases just some frames
less.

particle and struct runs at 75 fps on my machine and struct_array is maybe
70 fps.

But only rendering 1000 it shows a great difference. Here particle and
struct runs at ~1700 fps while struct_array only runs at 800 fps.

So when someone develops lets say struct_array as a first solution, spents
a lot of time to optimize it, to run twice as fast, like the other solutions
do. Then the result can be that this has no performance impact in
the final product.