Changed randi to be inclusive
Just makes more sense as randf is also inclusive. It somehow stupid to
say randi 1 10 and then it just returns numbers from 1 to 9. If i wanted
numbers from 1 to 9 i would have typed randi 0 9.

Obviously this is choosen so you can just pass the Length/Count of a
data-structure without calling -1 on it, but in my opinion a bad
language design decision.

In Perl you can do (0 .. @array) or (0 .. $#array) last one is
like "count-1", or in other words returns the last index instead
of the count.