Extended Drageable/getMouse to inlcude Camera
Up so far all small programs never used a camera. But when a camera
is included we need to convert screen positions to world Positions
to correctly check which element is draged in the world.

For this getMouse now can be optionally passed a Camera. When a Camera is
passed. Functions like processDrag can use the Camera to convert
screen coordinates to world position.