Spatial Tree
I don't like how the positions are calculated, seems to complicated to
me, but i stick with it at the moment and first make it somehow work.

Points are put into the tree and the regions are showing. The mouse
basically queries the Spatial Tree with its current position and draws
a line to the points found. Usually those results are than used for
collision detection/resolution.

Instead of just returning one chunk for a position than an implementation
should query a whole space that than returns all chunks inside that space.