Added Spatial Tree for Collision Detection
Performance now much better. Works fine with 5K-6K circles and still
over 60+fps and at 4 substeps. As Target FPS is set to 60 fps
it basically means it updates all circles at 240 hz.

During the substeps i just build the tree once and reuse it, even if to be
100% correct it must be rebuild after all updates are done. Performance
and accury can be tweaked by optimizing substeps and size of the Partial
Tree.

In a normal game we usually don't have 6000 circles stack on on top of each
other, so substeps could be much less. Also making the size of the chunks
smaller can improve performance. But it shouldn't be too small.

Completely depends on the game and the assets to set the correct size.