Changed VerletStructure & added point inside mesh function
For further addition the structure is changed. Instead of a list now
an array is picked. Before i never really used the structure anyway just
split it into its parts, but in the future the structure will stay intact.

An array allows efficent indexing that is needed for example for
pointInsideStructure function to check if a point is inside an concave
mesh.

Also already added the idea that some points generate a collision mesh
that is later used for collision detection/resolution.